New Colour Scheme because I felt the old one was too dark. Closes GH-5
This uses far lighter colours for a more 'github'y look. I have also removed the
unminified  source of  jQuery,  and added  in  some scripts  so  that the  pages
(should) render correctly in IE.

All    colour   declarations    in   the    css   have    been   extracted    to
`public/src/colours.css` so  that it  is easier  for someone  to make  their own
colour scheme.

Please do take the time to play  with this in internet explorer, because I don't
use a mac so i  can't test in any of the IE versions. Just  open a new ticket or
add a comment to an open one if there are any problems. Please Include a link to
a screenshot so  that we can see what's wrong.  The Perfect Renderings (provided
by Safari 4) can be seen here (the first four images):

http://www.flickr.com/photos/24295828@N03/sets/72157618065085813/